Use an interface for the factory and move it to the correct directory.
