+++ tests/RichEditorServiceProvider.php
@@ -4,9 +4,9 @@ declare(strict_types=1);
 
 namespace Tests\MonsieurBiz\SyliusRichEditorPlugin;
 
+use MonsieurBiz\SyliusRichEditorPlugin\Factory\UiElementFactory;
 use MonsieurBiz\SyliusRichEditorPlugin\Twig\RichEditorExtension;
 use Symfony\Bundle\FrameworkBundle\Test\KernelTestCase;
-use MonsieurBiz\SyliusRichEditorPlugin\UiElement\Factory;
 use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 use Twig\Environment;
 
@@ -17,8 +17,8 @@ final class RichEditorServiceProvider extends KernelTestCase
         self::bootKernel();
         $container = self::$container;
 
-        /** @var Factory $factory */
-        $factory = $container->get(Factory::class);
+        /** @var UiElementFactory $factory */
+        $factory = $container->get(UiElementFactory::class);
         /** @var Environment $environment */
         $environment = $container->get(Environment::class);
         /** @var EventDispatcherInterface $eventDispatcher */
